Add an option for East Asian Ambiguous character width
East Asian Ambiguous characters are traditionally treated as two column width
characters, and many fonts are designed to display them as so. There are two
types of applications that regard this width. One is terminal emulators like
hterm, and the other is applications that use libcurses. The mismatch of the
column width between these applications causes an odd rendering. Adding an
option solves this issue.

In hterm, it has a JavaScript port of wcwidth. Even though the original version
of wcwidth includes the treatment for East Asian Ambiguous characters, the
ported version omits it. This commit add this missing part and add an option to
switch the width.

Change-Id: I14b7c052202203691e57043e10bab2b2a83de51d
Reviewed-on: https://chromium-review.googlesource.com/202247
Reviewed-by: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>