libdot: Fix rgbToHex for red values lower than 16.
To observe the affected behavior, try changing default foreground,
background, or cursor color in the Secure Shell preference page to
something with a red value below 16. The color will be assumed to
be black.

For example, without this fix, setting the cursor color to
rgb(0, 255, 255) results in the following warning in Chrome's console
and the color changing to black:

--- cut here ---
The specified value '#0' does not conform to the required format.
The format is '#rrggbb' where rr, gg, bb are two-digit hexadecimal
numbers.
--- cut here ---

Change-Id: I5978e00c152720ae3f8d4e04b404fda7d3b262e4
Reviewed-on: https://chromium-review.googlesource.com/260069
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Rob Ginda <rginda@chromium.org>