libdot: mkzip: split channel & version promotion
Change the default promotion behavior to only promote the channel.
This way we can do releases more like Chrome OS where a version is
channel agnostic, and when we move the dev to stable, the only thing
that changes in the channel info.  This makes it easier to look up
in the ChangeLog history too.

We unify the init from dir & zip code paths so they behave the same
rather than forcing the zip to always promote.

Change-Id: Ic9ea9d53964bc8085575ae0630949ff6221acec1
Reviewed-on: https://chromium-review.googlesource.com/491626
Reviewed-by: Brandon Gilmore <varz@google.com>
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>