libdot: add helper for creating unique enums
Most JS linters warn when you try to use the String class (because people
should use string primitives everywhere), so our usage to create enums
makes the linters unhappy.  We can't use JSDoc suppress tags because they
have file and function level scope, but we need variable/one-line scope.

Create a dedicated function to wrap this so we can document the usage
clearly, and so we can suppress the lint checkers in this one place.

Change-Id: Id406224ed4c07a5a258ad4b9ef1e9caee6b73142
Reviewed-on: https://chromium-review.googlesource.com/508413
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>