libdot: wc: include leading combining characters
The substring logic will skip any leading combining characters because
their display width is zero.  This is what we expect, but it's also not
exactly something people should be doing as it doesn't make sense: this
API is meant for splitting strings based on their display which is based
on graphemes.

However, there is an edge case that can come up when dealing with random
user generated content: the start of the string might contain combining
characters.  Since there aren't any codepoints that they'd be attached
to, we skip them all.  We can special case this one situation though as
it should also be a common one and thus (hopefully) be a general slight
improvement in general.

BUG=chromium:212707

Change-Id: I4690e9f1b597a1fdcea87a478ef0971ef19103e5
Reviewed-on: https://chromium-review.googlesource.com/624915
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>