libdot: wc: update to Unicode 10.0.0
We first have to create a script to programatically generate the lookup
tables.  Markus Kuhn's uniset table works for combining characters, but
it has a hardcoded set for the unambiguous wide characters, and requires
a non-standard preformatted text file for the ambiguous wide characters.
So rather than continuing to patch those pieces together by hand (which
is pretty error prone), write a new python helper that works off of all
the official Unicode database files.  Also, uniset is perl ;).

With that in hand, we can document the procedure for merging Unicode
updates, and then we follow it to import the 10.0.0 release.  There's a
lot of changes here because the last time we updated was Unicode 5.0.0
from ~11 years ago, and because there's a bunch of new single codepoints
sprinkled in different blocks.  The astute reader might also notice that
the previous unambiguous table included some codepoints that this update
does not, but that is by design: the old version included unallocated
blocks under the assumption they'd all be wide in the future (which ended
up not being the case).

Since we're basically throwing out the old tables with completely new,
take the opportunity to reformat to better align with our style guides.

The tests take a hit because we were using an emoji codepoint that under
the older standard had a width of 1, but the new one sets it to 2.

The new helper was sanity checked by running it against the new and old
releases and comparing them against the output of Markus's uniset tool.
Furthermore, many of the new single elements that showed up were spot
checked against the Unicode database to make sure they should be there.

Change-Id: Ifab96c8531be6f8ca927d13f2faebe2c9f1e3001
Reviewed-on: https://chromium-review.googlesource.com/646797
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>