libdot: wc: move east asian chars into a lookup table
The original implementation hand unrolled the codepoint search because
the number of matching ranges was "low" in Unicode 5.0.0 (about 11).
However, in the current release, the number of ranges has gone up quite
a bit, so we'll want to switch to a table lookup.

Before we actually update to a newer Unicode release, convert the
existing code to use a table lookup.  There should be no functional
difference here and should make future reviews easier.

Change-Id: I5f00763b9f87bbc56a81e3816cff1566283b2642
Reviewed-on: https://chromium-review.googlesource.com/646796
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>