libdot: retry storage saves when quota is exceeded
When restoring a backup with a lot of entries (>100), we end up
exceeding the Chrome sync storage write quota and all following
writes get rejected.  Handle this by situation by retrying the
write once a second until it passes.

Since we are given callbacks by the higher levels, we can still
properly notify interested parties when it finally succeeds.

In practice, if you were to restore a large backup, you'll have
to wait for a minute for the quota to become available again.
We don't currently have any UI to notify the user, but we'll
wait to do that later.  Lets just make it work now.

BUG=chromium:764759

Change-Id: I2bd53a52c3aae143504d92dcf159bc04bdce5981
Reviewed-on: https://chromium-review.googlesource.com/665669
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brandon Gilmore <varz@google.com>