libdot: coalesce writes when importing json files
When importing many children elements, we end up rewriting the array
list many times and committing each update to storage.  This in turn
eats up our quota and cause use to overrun it when there are many
active profiles.

Change the json import logic so that we suppress the intermediate and
do a single flush at the end.

Bug: chromium:764759
Change-Id: I8fe0725ccfc006f2d269a88cb9bb4d672c4da929
Reviewed-on: https://chromium-review.googlesource.com/c/665670
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>