nassh: support running as an extension
We largely don't need to change anything here.  Just a lot of
documentation updates.

We tweak the nassh code to support the browserAction API.  This is
activated when the icon is clicked.  We use it to launch a window.
In addition, we set the default popup to list existing connections.
The UI here is likely to change, so just putting together something
simple and reasonable for users to start experimenting with.

Finally we tweak the manifests to generate multiple packages.  The
main manifest json is a hand maintained symlink to help with local
testing.

We start the extension version off low so we can do some early
experimentation before getting it onto the same release track.

We require Chrome 65+ as those are the first to whitelist us.

Change-Id: I7c0c3e8960b3ba31c15c8bf4c06fccb29a3b7a5a
Reviewed-on: https://chromium-review.googlesource.com/832675
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brandon Gilmore <varz@google.com>