libdot/nassh: shuffle plugin paths for CWS filtering
The CWS supports smaller downloads when using NaCl plugins by only
sending nexe's that match the client's architecture, but it requires
the files be placed in specific subdirs.  Add a helper script that
rewrites any NaCl manifests it finds to the right layout so we don't
have to do file shuffling by hand all the time.

Change-Id: Ic4209a53bc98ae7e3297ec49edcd83c66a44ec8d
Reviewed-on: https://chromium-review.googlesource.com/843816
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>