libdot: i18n: move getMessage helper here
Lets move this out of the message manager so it can be more widely
used, and benefit from the chrome/browser logic in lib.i18n.

This changes the behavior of lib.MessageManager.get slightly in that
we let lib.i18n.getMessage do all the argument replacing for us.
This in turn fixes some latent bugs when getting translations that
the browser knows about -- if the replacements aren't specified, the
browser also strips out the $# values.

BUG=chromium:857538

Change-Id: I2db519d3ceb857419aceb26b315095e453e5efa7
Reviewed-on: https://chromium-review.googlesource.com/1135781
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>