libdot: convert concat to a python script
The current concat.sh script is getting a bit unmanageable and slow:
it takes 5-6 seconds to run on nassh.  Replace it with a Python script
that takes ~300msec.

We don't support the forever/inotify logic as it doesn't seem to be
used that much and it's a pain to convert.

Now that we have a more powerful language, we can replace most of the
shell usage in the concat files with native operations.  Namely, we
add direct support for date, base64, and for changelog parsing.

Change-Id: I99e82742d8a0a45a5d868dfcace6e2c2e8b99f74
Reviewed-on: https://chromium-review.googlesource.com/1156073
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>