libdot: codec: new module for binary/UTF8/UTF16 helpers
We have a few helpers for dealing with converting between different
encodings spread across the projects.  Some have better names than
others, and some (like lib.utf8) would be replaced entirely by newer
web standards.

Lets start a new home for all the helpers that'd still exist when
we drop lib.utf8, and lets aim for clearer names from the start.

This seeds the module with funcs from nassh.Stream.  We leave the
base64 logic in nassh.Stream as that largely exists only for the JS
and ppapi plugin to communicate, and ideally we'll convert all of
that to arrays at some point.

Bug: 776106
Change-Id: I8f16cdf4c79e18a48b21cb3d95d978c8ba0296c5
Reviewed-on: https://chromium-review.googlesource.com/c/1391808
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>