hterm: optimize image display with Blobs & ArrayBuffers
Currently we display the base64 data coming straight from the user via
data URLs.  This means we stuff the string directly into the DOM which
Chrome (currently) does not handle well when the image is large (>1MB).
If we turn it into an ArrayBuffer though, we can pass it via a Blob to
URL.createObjectURL, and Chrome is able to efficiently display things.

This also has the nice side-effect of fixing non-inline file transfers
in more cases.  Chrome, when given a data URL, would download the base64
data as base64 strings.  The user could convert it by hand, but it's a
bit of a pain depending on the OS.  By using URL.createObjectURL, Chrome
will save it with the correct bytes all the time.

This flow too avoids the ~2MB limit that Chrome imposes on data URLs.
Blobs & URL.createObjectURL allow for (basically) unlimited file sizes.

Change-Id: Idadbdc5d19d681b4ddbdf48b71671fca9c407d83
Reviewed-on: https://chromium-review.googlesource.com/c/1408095
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>