libdot: npm: add more test/dev packages
chai is an assert test library which allows us to maintain the same
API across test frameworks.  It's a small library and allows us to
migration in smaller steps.

mocha is the new test framework so we can convert our unittests over
to running from the command line (so we can run them in the CI).  We
only convert the in-browser test suites for now as we experiment with
the new setup.

jsdom provides a browser-like environment (e.g. the Window object)
when running in headless tests.  This allows us to run more browser
code in the headless node environment.

uglify-js only supports ES5, so in practice it doesn't work for a
lot of our code.  Switch to terser which supports ES6+ (and because
the uglify-es fork is deprecated).

firebase tools are for managing the hterm.org website.

Require a newer csso-cli version to include fixed source map names.

Change-Id: I7d9aa05b3e965ab9d74a188fb1cd10f370d9cabf
Reviewed-on: https://chromium-review.googlesource.com/c/1476660
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>