libdot: codec: add performance notes wrt String.fromCharCode.apply
For small arrays, String.fromCharCode.apply can be faster than this
for loop implementation.  However, it's implemented as a recursive
call, so it'll fail with large inputs (200k+).

Further, running this code frequently shows that, while the apply
method might be faster if called a few times (<10), when called a
lot in higher level loops (>100), its performance is quite variable.
The for loop code on the other hands tends to be pretty stable.

If we could assume the input was an array buffer, we could utilize
typed arrays to use String.fromCharCode.apply on smaller chunks (so
we wouldn't blow the stack).  But even this hybrid approach has the
variablity issue.  Something like:
  const chunk = 10000;
  for (let i = 0; i < array.length; i += chunk)
    ret += String.fromCharCode.apply(null, array.subarray(i, i + chunk));

Bug: chromium:924764
Change-Id: Ie3510b4d2f55ec5724cf0e38d7609e00ee7d5598
Reviewed-on: https://chromium-review.googlesource.com/c/1437527
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>