libdot: concat: improve plain text resources
When loading files directly, we have to make sure they're in UTF-8
encoding before writing to the output.  We've been doing this only
for base64 outputs, but we haven't noticed for others because we
haven't been using them before (we will in a follow up commit).

We can also improve handling of plain text resources by not munging
the whitespace that might be embedded in the file (which is what our
textwrap module does currently).  Auto-escape all those characters.

Change-Id: I190655df8de10bc05d7c4cb25e8556b5a87f12f6
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1646582
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>