tests: leverage mocha-headless-chrome for headless/CI
We now support running unittests via headless chrome by default.
We pin the version to test against & download it on the fly so we
can run the tests in our CI.

The default load_tests behavior is now to use the headless mode.
The new --visible flag can be used to get the old behavior where
we run the browser in interactive mode.

Change-Id: If10858f55f16dc69d42826f9ff2308f264b9bf93
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1712314
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>