libdot: load_tests: add a --no-sandbox flag
Running Chrome requires a working sandbox normally.  This works on
systems with userns enabled, but if that's disabled/unavailable, then
the tests fail immediately with sandbox errors.  Add a --no-sandbox
flag so people can still run tests albeit w/out sandboxing enabled.
For our needs, this shouldn't be a problem as we are only running our
own code.

Change-Id: Iccd65fcfb3a2480f43bcd5ae872ea3651130be54
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1727433
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>