libdot: storage: create common interface
Create a common lib.Storage interface for the
lib.Storage.Local, lib.Storage.Memory, and
lib.Storage.Chrome classes to implement.

Changed semantics of getItems in all
implementation s to return all values if null
is passed in.  This is the way
lib.Storage.Chrome works, and it makes
sense for the others to be similar.

Bug: 998928
Change-Id: I0c862ccc7669daf4789004e9f42abcc85a997849
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1816019
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>