libdot: lint: improve gerrit comments output
Parser errors won't include a ruleId so don't attempt to link to the
eslint documentation as it'll be invalid.  Which means it might not
be clear the source of the message, so add a [eslint] prefix to all
warnings coming from it.

Tweak the column handling as gerrit appears to count from 0 (the left
edge of the selection) while eslint counts from 1.

Also don't automatically set the end line to the current line + 1.
If eslint doesn't produce an endLine setting, then it's probably all
contained in the current line.

If running by hand with a relative path like --gerrit-comments-file
foo.json, don't try to create the directory "" as it'll fail.

Finally, it seems that not all Python versions (like 3.5) handle
passing binary data into json.loads.  Convert it to UTF-8 first.

Change-Id: I35728b578f69cd5a6acd87608333f6e9a2a4b85f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1820900
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>