terminal: remove checkTypes suppression
It doesn't seem possible to get closure compiler
to successfully read the nassh_deps.rollup.js file
and as a result, calls to load modules from it
fail.  The fix for this, is to leave it, and
lit_element.js out of compilation, and use
externs/lit_element.js and ignore moduleLoad errors.

Change-Id: If4da67de9e762d411dc4ce4a905afe72dea8aec2
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1855343
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Julian Watson <juwa@google.com>