use UTF-8 encoding in more places
We're seeing the generated hterm concat files with corrupted encodings.
Update all the places where we read/write files in Python to use an
explicit encoding so we don't rely on the ambient runtime.  Also update
the kokoro settings to force default LANG to C.UTF-8 when running in
the Debian container.

Change-Id: Ife2e1e34599d1396692ae444c4cbf46079d63a48
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2020760
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>