libdot: Lookup en directly rather than en_US for messages
Our messages files use 'en' as the default for 'en_US'
and so the first attempt to find 'en_US' always fails
with js console errors before we then try 'en' as
the default.

This change explicitly swaps en_US to be en so that
we get the correct messages.json file first time
without errors.

Change-Id: I9ee67612541281e1379e45097c37945a4ae39fa6
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2041932
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>