nassh: omnibox: workaround bugs in incognito mode
Some users want to run the extension in incognito mode, so enable
that.  The app has defaulted to incognito=split due to it being a
legacy packaged app, but the extension defaults to spanning.  Set
both to split explicitly to avoid confusion.

With that in place, we find R84 has some bugs wrt curentWindow and
chrome.tabs APIs in incognito mode.  We can workaround it by using
the window id ourselves, so do so until Chrome can get fixed.

Bug: chromium:1075427
Test: checked omnibox behavior in normal & incognito modes after
      allowing the extension to be used in incognito

Change-Id: I94bf4a14506b3046afc2a3748ca3166c781fb219
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2172621
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>