nassh: fsp: switch to Chrome error constants
This is to prepare for moving to the new FSP externs.  We should be
using these error enums anyways instead of ad-hoc strings.

We also have to add some Chrome mocks for tests.  This is enough to
unblock the current usage, but we should look at some projects more
complete eventually.

Change-Id: Ia124e9e68e7e362eaec8fe85aee42ed867b598e7
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2212150
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>