libdot: storage: chrome: fix getItem API
All other implementations invoke the callback with the value instead
of with an object with the value (or undefined if the key isn't set).
Change the chrome API to match.

This shouldn't cause issues are there aren't any users currently of
the getItem API when using the chrome storage implementation.

Change-Id: I6e4ae3ab97da522330c527cfa338af883eb4a00b
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2257004
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>